board 타입 long 으로 변경

+++ src/main/java/org/zerock/domain/ReplyVo.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class ReplyVo {
 	private int REPLY_NO;
-	private int BOARD_NO;
+	private long BOARD_NO;
 	private int MEM_NO;
 	private String REPLY_CONTENT;
 	private Date REPLY_DATE;
